[#35] :truck: rename: 일기장 diary -> book 이름 변경
