--- client/src/recoil/new-diary/index.js
@@ -1,1 +1,0 @@
-export { selectedColorAtom, diaryTitleAtom } from "./atom";
